:white_check_mark: Get coverage back up for all work done
