+++ example/counter/layout/not_default.html
@@ -1,7 +1,8 @@
 <html>
   <head>
+    <title>Different Layout</title>
   </head>
   <body>
-    { content }
+    <<<RIOT>>>
   <body/>
 </html>
