+++ example/counter/server.js
@@ -22,7 +22,7 @@ server.register([Vision, Inert], function (err) {
     relativeTo: __dirname,
     compileOptions: {
       layoutPath: Path.join(__dirname, 'layout'),
-      layout: true
+      layout: 'not_default.html'
     },
     path: 'views'
   });
