+++ example/counter/views/index.tag
