+++ example/new/layout/not_default.html
@@ -1,0 +1,7 @@
+<html>
+  <head>
+  </head>
+  <body>
+    <<<RIOT>>>
+  <body/>
+</html>
