+++ lib/index.js
@@ -46,11 +46,9 @@ function loadLayout (filePath) {
  * that changes made in components are re-loaded
  */
 function unCache () {
-  if (Array.isArray(tagsLoaded)) {
-    tagsLoaded.forEach(function (file) {
-      delete require.cache[require.resolve(file)];
-    });
-  }
+  tagsLoaded.forEach(function (file) {
+    delete require.cache[require.resolve(file)];
+  });
   tagsLoaded = null;
   layoutCache = {};
 }
@@ -74,14 +72,6 @@ function compile (template, compileOpts) {
     + '\n<script src="/compiled.js" ></script>'
     + '\n<script>\nriot.mount("*",' + JSON.stringify(context) + ')\n</script>'
     ;
-    /*
-     * Delete the view and layout Tags from the require cache
-     * so we don't need to restart the app to see view changes.
-     * Skipped By default when NODE_ENV=production`.
-     */
-    if (mergedOpts.removeCache) {
-      unCache();
-    }
 
     if (mergedOpts.layout) {
       layoutFileName = mergedOpts.layout === true
@@ -90,10 +80,19 @@ function compile (template, compileOpts) {
       ;
       layoutFilePath = Path.join(mergedOpts.layoutPath, layoutFileName);
 
-      return loadLayout(layoutFilePath).replace('<<<RIOT>>>', content);
+      content = loadLayout(layoutFilePath).replace('<<<RIOT>>>', content);
     }
 
-    return content;
+    /*
+     * Delete the view and layout html from the require cache
+     * so we don't need to restart the app to see view changes.
+     * Skipped By default when NODE_ENV=production`.
+     */
+    if (mergedOpts.removeCache) {
+      unCache();
+    }
+
+    return '<!DOCTYPE html>' + content;
   };
 }
