+++ test/simulate_prod.test.js
@@ -6,20 +6,20 @@ var Hapi = require('hapi');
 var Vision = require('vision');
 var assert = require('assert');
 
-var HapiRiot = require('../lib/index.js');
 
 test('Simulate Production', function (t) {
   var server = new Hapi.Server();
 
   server.connection(0);
 
+  decache('../lib/index.js');
+
   server.register(Vision, function (err) {
-    // console.log('ERR >>>> ', err);
     assert(!err); // Halt start if Vision fails to load.
-    decache('../lib/index.js');
+    t.plan(2);
 
     server.views({
-      engines: { tag: HapiRiot },
+      engines: { tag: require('../lib/index.js') }, //eslint-disable-line
       relativeTo: __dirname,
       path: 'views',
       compileOptions: { removeCache: false } // i.e. cache all views
@@ -40,7 +40,15 @@ test('Simulate Production', function (t) {
         response.result.indexOf(expected) > -1,
         'Result contains ' + expected
       );
-      server.stop(t.end);
+    });
+
+    server.inject({ url: '/' }, function (response) {
+      var expected = 'My Amazing Title!';
+
+      t.ok(
+        response.result.indexOf(expected) > -1,
+        'Uses cache ok'
+      );
     });
   });
 });
